requirements/local.txt
-r base.txt
django-debug-toolbar==3.4.*
gunicorn==20.1.*
whitenoise==6.1.*
django-mptt==0.13.*
mock==4.0.*